requirements.txt
gunicorn
flask
flask-restful
pytesseract
tesseract
Pillow